Using the agent instance to get the name
