+++ deep_quoridor/src/train_sb3.py
@@ -14,11 +14,10 @@ TODO:
 import glob
 import os
 import time
-from dataclasses import asdict
 
-import pettingzoo.utils
 import quoridor_env
 import torch
+from agents.sb3_ppo import SB3ActionMaskWrapper
 from gymnasium import spaces
 from sb3_contrib import MaskablePPO
 from sb3_contrib.common.maskable.policies import MaskableActorCriticPolicy
@@ -26,6 +25,7 @@ from sb3_contrib.common.wrappers import ActionMasker
 from stable_baselines3.common.torch_layers import FlattenExtractor
 
 import wandb
+from deep_quoridor.src.agents.sb3_ppo import SB3PPOAgent
 from wandb.integration.sb3 import WandbCallback
 
 
@@ -46,85 +46,11 @@ class DictFlattenExtractor(FlattenExtractor):
         return thobs
 
 
-class SB3ActionMaskWrapper(pettingzoo.utils.BaseWrapper):
-    """
-    Wrapper to allow PettingZoo environments to be used with SB3 illegal action masking.
-    In particular it adapts PettingZoo, since the Action Masking part of it is already implemented
-    in SB3_contrib as the MaskablePPO.
-    The required changes are minor:
-    - Present observation_space and action_space as props instead of methods
-    - return last() on step()
-    - return observation on reset()
-    - return only the observation (not the action mask) in observe()
-    - provide a method to get to the action mask
-    """
-
-    def __init__(self, env, rewards_multiplier: float = 1000):
-        super().__init__(env)
-        self.rewards_multiplier = rewards_multiplier
-
-    def reset(self, seed=None, options=None):
-        """Gymnasium-like reset function which assigns obs/action spaces to be the same for each agent.
-
-        This is required as SB3 is designed for single-agent RL and doesn't expect obs/action spaces to be functions
-        """
-        super().reset(seed, options)
-
-        # SB3 needs observation and action spaces as props instead of methods (?)
-        full_observation = super().observation_space(self.possible_agents[0])
-        self.observation_space = full_observation["observation"]
-        self.action_space = super().action_space(self.possible_agents[0])
-
-        # Return initial observation, info (PettingZoo AEC envs do not by default)
-        return self.observe(self.agent_selection), {}
-
-    def step(self, action):
-        """Gymnasium-like step function, returning observation, reward, termination, truncation, info.
-
-        The observation is for the next agent (used to determine the next action), while the remaining
-        items are for the agent that just acted (used to understand what just happened).
-        """
-        current_agent = self.agent_selection
-
-        super().step(action)
-
-        return (
-            self.observe(current_agent),
-            self.rewards[current_agent] * self.rewards_multiplier,
-            self.terminations[current_agent],
-            self.truncations[current_agent],
-            self.infos[current_agent],
-        )
-
-    def observe(self, agent):
-        """Return only raw observation, removing action mask."""
-        obs = super().observe(agent)["observation"]
-        # # Take obs, which is a dict with some arrays, and convert it to a flat numpy array
-        # return np.concatenate(
-        #     [
-        #         obs["board"].flatten(),
-        #         obs["walls"][0].flatten(),
-        #         obs["walls"][1].flatten(),
-        #         [obs["my_walls_remaining"], obs["opponent_walls_remaining"]],
-        #     ]
-        # )
-        return obs
-
-    def action_mask(self):
-        """Separate function used in order to access the action mask."""
-        return super().observe(self.agent_selection)["action_mask"]
-
-
 def mask_fn(env):
     # Do whatever you'd like in this function to return the action mask
     # for the current env. In this example, we assume the env has a
     # helpful method we can rely on.
     return env.action_mask()
-
-
-def make_model_id(env_kwargs):
-    # HACK: This is takenfrom trainable_agent model_id()
-    return f"sb3ppo_B{env_kwargs['board_size']}W{env_kwargs['max_walls']}_mv0"
 
 
 def train_action_mask(env_fn, steps=10_000, seed=0, **env_kwargs):
@@ -153,7 +79,7 @@ def train_action_mask(env_fn, steps=10_000, seed=0, **env_kwargs):
         callback=WandbCallback(gradient_save_freq=1000),
     )
 
-    model_id = make_model_id(env_kwargs)
+    model_id = SB3PPOAgent(**env_kwargs).model_id()
     local_filename = f"{model_id}_{time.strftime('%Y%m%d-%H%M%S')}.zip"
     model.save(local_filename)
     artifact = wandb.Artifact(f"{model_id}", type="model")
@@ -176,7 +102,7 @@ def eval_action_mask(env_fn, num_games=100, render_mode=None, **env_kwargs):
     print(f"Starting evaluation vs a random agent. Trained agent will play as {env.possible_agents[1]}.")
 
     try:
-        model_id = make_model_id(env_kwargs)
+        model_id = SB3PPOAgent(**env_kwargs).model_id()
         latest_policy = max(glob.glob(f"{model_id}*.zip"), key=os.path.getctime)
     except ValueError:
         print("Policy not found.")
